third_party/WebKit/LayoutTests/fast/multicol/out-of-flow/offset-properties.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<div style="position:relative; border:1px solid;">
    <div style="columns:3; column-gap:0; column-fill:auto; margin:100px; width:300px; height:50px;">
        <div style="height:70px;"></div>
        <div id="elm" style="position:absolute;"></div>
    </div>
</div>
<script>
    test(function() {
        assert_equals(document.getElementById('elm').offsetLeft, 200);
        assert_equals(document.getElementById('elm').offsetTop, 120);
    }, "auto-positioned abspos, containing block on the outside of the multicol.");
</script>
